Adds the capability to set a user to forcely change its own password at next LogOn
Force user to not reuse last n passwords

Moves IPasswordService implementation to Orchard Users
Creates Extensions to share management of Password operations across Services

Some refactoring

Password History Policy:
- New User Evente (ChangingPassword)
- Settings to enable the policy
- Security service interfaces to abstract history management
- User service implementations to concretely manage history

PasswordHistoryPolicy:
- keep in count the password stored within the UserPart as a not reusable password

WIP automated suspension of inactive users

Disable users that have been inactive for longer than a specified number of
days, except when they are SiteOwner, or they have a specific flag set to
prevent their suspension.

Provider to prevent suspension of users based on assigned roles

cleanup. Refactor of migrations.

Added action to ask for the challenge email to be resent.
Challenge email is sent again if a user tries to register anew with an email
address they had used to create an account earlier if the email address isn't
validated yet.

During registration, if a user inserts the information of an existing account
and that account should still validate its email address, the user is presented
a link to request a new challenge email to be sent.

Added a link to the action to request a new challenge email in the case when
the nonce fails to validate.